backend/pyproject.toml
[project]
name = "resume-optimizer"
version = "0.1.0"
description = "AI-powered resume optimization pipeline using three sequential agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "openai>=1.12.0", # For OpenRouter API compatibility
    "python-dotenv>=1.0.0",
    "plotly>=5.19.0",
    "pandas>=2.2.0",
    "python-docx>=1.1.0",
    "pillow>=10.2.0",
    "httpx>=0.27.0",
    "tiktoken>=0.6.0", # For token counting
    "jinja2>=3.1.3", # For HTML templating
    "beautifulsoup4>=4.12.2", # For HTML -> DOCX parsing
    "weasyprint>=61.2", # Optional: PDF generation (not required at runtime)
    "docxtpl>=0.16.7", # For DOCX template filling
    "exa_py>=1.0.15", # For live job posting retrieval
    "PyGithub==2.8.1", # For GitHub API access
    "xhtml2pdf==0.2.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]
